+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Nothing yet.
+
+## [1.5.0] - 2023-08-25
+
 ### Added
 
 - Implemented user password change.
+- Added claim constants.
 
 ### Changed
 
@@ -68,7 +73,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implemented StringExtensions.
 
-[unreleased]: https://github.com/Logitar/Logitar.NET/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/Logitar/Logitar.NET/compare/v1.5.0...HEAD
+[1.4.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Logitar/Logitar.NET/releases/tag/v1.1.0...v1.2.0